% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIntervals.R
\name{plotIntervals}
\alias{plotIntervals}
\title{Plot intervals}
\usage{
plotIntervals(intervals)
}
\arguments{
\item{intervals}{data frame (see \link{generateIntervals} for
the required format)}
}
\value{
ggplot object
}
\description{
Generates a plot of a set of intervals.  This is intended for simple
visualizations and does not offer any degree of customization.
}
\examples{
intervals <- generateIntervals(10)
p <- plotIntervals(intervals)
p
}
